Moving calculateWinner to utils.js
