+++ react/training/src/redux/game/reducers.js
@@ -1,21 +1,11 @@
 import { actionTypes } from './actions';
+import calculateWinner from './utils';
 
 const initialState = {
   history: [{ squares: Array(9).fill(null) }],
   stepNumber: 0,
   xIsNext: true,
   winner: ''
-};
-
-const calculateWinner = squares => {
-  const lines = [[0, 1, 2], [3, 4, 5], [6, 7, 8], [0, 3, 6], [1, 4, 7], [2, 5, 8], [0, 4, 8], [2, 4, 6]];
-  for (let i = 0; i < lines.length; i++) {
-    const [a, b, c] = lines[i];
-    if (squares[a] && squares[a] === squares[b] && squares[a] === squares[c]) {
-      return squares[a];
-    }
-  }
-  return null;
 };
 
 export default function reducer(state = initialState, action) {
